Allow reading from *HUGE* files.
In contrast to the previous solution, when a file was read entirely and
profiles were processed profile per profile, now the file is read
in buffers of some relatively small quantity (configurable). This
buffer, or if it is not big enough, just as much as needed is read and
processed. Processing a several GB file will then consume virtually no
RAM. What a deal.